// old src/types/options.d.ts
export interface DiscordOptions {
    shards?: number | number[] | 'auto';
    closeTimeout?: number;
    shardCount?: number;
    makeCache?: CacheFactory;
    allowedMentions?: MessageMentionOptions;
    partials?: Array<Partials>;
    failIfNotExists?: boolean;
    presence?: PresenceData;
    intents: number[];
    waitGuildTimeout?: number;
    sweepers?: SweeperOptions;
    ws?: WebSocketOptions;
    rest?: RESTOptions;
    jsonTransformer?: (obj: unknown) => unknown;
}
export interface ClientOptions {
    ownerId?: string | string[];
    permissions?: {
        client?: Array<string>;
        user?: Array<string>;
        dm?: boolean;
    };
    ignore?: {
        owner?: boolean;
        users: Array<string>;
        channels: Array<string>;
        guilds: Array<string>;
    };
    log?: {
        ready: boolean;
        error: boolean;
    };
    cooldown?: {
        command?: number;
        cache?: number;
    };
}
export interface HandlerOptions {
    message?: MessageCommandHandlerOptions;
    slash?: SlashCommandHandlerOptions;
    context?: ContextCommandHandlerOptions;
    events?: EventHandlerOptions;
    inhibitors?: InhibitorHandlerOptions;
    webhook?: WebhookOptions;
}
export interface ExtensionOptions {
    extensions: Array<XernerxExtensionBuilder>;
    logging?: boolean;
}
// ...
export interface MessageCommandHandlerOptions {
    directory: string;
    prefix: string[] | string;
    allowMention?: boolean;
    cooldown?: number;
    userPermissions?: bigint[];
    clientPermissions?: bigint[];
    handleEdits?: boolean;
    handleDeletes?: boolean;
    handleTyping?: boolean;
    util?: boolean;
    logging?: boolean;
}
export interface SlashCommandHandlerOptions {
    directory: string;
    guildId: string;
    global: boolean;
    cooldown?: number;
    userPermissions?: bigint[];
    clientPermissions?: bigint[];
    util?: boolean;
    logging?: boolean;
    defer?: {
        reply?: boolean;
        ephemeral?: boolean;
        fetchReply?: boolean;
    };
}
export interface ContextCommandHandlerOptions {
    directory: string;
    guildId: string;
    global: boolean;
    cooldown?: number;
    userPermissions?: bigint[];
    clientPermissions?: bigint[];
    util?: boolean;
    logging?: boolean;
    defer?: {
        reply?: boolean;
        ephemeral?: boolean;
        fetchReply?: boolean;
    };
}
export interface EventHandlerOptions {
    directory: string;
    logging?: boolean;
}
export interface EventOptions {
    name: string;
    emitter?: EventEmitterType | string;
    type?: 'discord' | string;
    once?: boolean;
}
export interface InhibitorHandlerOptions {
    directory: string;
    logging?: boolean;
}
export interface InhibitorOptions {
    name: string;
    type: InhibitorType;
}
export interface WebhookOptions {
    topgg: {
        token: string;
    };
    dbl: {
        token: string;
    };
    config: {
        emits: boolean;
        logs: boolean;
    };
}